+++ Owen013.MovementMod/Config.cs
@@ -73,7 +73,7 @@ public static class Config
 
     public static float FloatyPhysicsMinGravity { get; private set; }
 
-    public static float MaxWallJumps { get; private set; }
+    public static int MaxWallJumps { get; private set; }
 
     public delegate void ConfigureEvent();
 
@@ -115,7 +115,7 @@ public static class Config
         FloatyPhysicsMinAccel = config.GetSettingsValue<float>("Floaty Physics Minimum Acceleration");
         FloatyPhysicsMaxGravity = config.GetSettingsValue<float>("Floaty Physics Minimum Gravity");
         FloatyPhysicsMinGravity = config.GetSettingsValue<float>("Minimum Gravity");
-        MaxWallJumps = config.GetSettingsValue<float>("Maximum Number of Wall Jumps");
+        MaxWallJumps = config.GetSettingsValue<int>("Maximum Number of Wall Jumps");
 
         if (ModMain.Instance.SmolHatchlingAPI != null && ModMain.Instance.SmolHatchlingAPI.UseScaledPlayerAttributes())
         {
